+++ packages/e2e-playwright/fixtures/card.fixture.ts
@@ -1,5 +1,4 @@
-import { test as base, expect } from '@playwright/test';
-import { URL_MAP } from './URL_MAP';
+import { test as base, expect } from './base-fixture';
 import { Card } from '../models/card';
 import { BCMC } from '../models/bcmc';
 import { CardWithAvs } from '../models/card-avs';