+++ packages/lib/eslint.config.js
@@ -14,7 +14,17 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
 const config = tseslint.config(
     {
         name: 'Global ignore',
-        ignores: ['storybook-static/*', 'coverage/*', 'dist/*', 'config/*', 'auto/*', 'postcss.config.cjs', 'eslint.config.js', '**/*_*.*']
+        ignores: [
+            'storybook-static/*',
+            'coverage/*',
+            'scripts/*',
+            'dist/*',
+            'config/*',
+            'auto/*',
+            'postcss.config.cjs',
+            'eslint.config.js',
+            '**/*_*.*'
+        ]
     },
     eslint.configs.recommended,
     jsxA11y.flatConfigs.strict,