+++ packages/lib/src/components/Card/components/CardInput/components/StoredCardFields.test.tsx
@@ -36,7 +36,7 @@ describe('StoredCard', () => {
 
         // Look for cvc field elements
         expect(screen.getAllByText('Security code', { exact: true })).toBeTruthy();
-        expect(screen.getByRole('img', { name: 'Security code' })).toBeTruthy();
+        expect(screen.getByRole('img', { name: 'Security code 3 digits on back of card' })).toBeTruthy();
     });
 
     test('Renders a StoredCard field, without expiryDate (relevant data is null); and with cvc field', () => {
@@ -49,7 +49,7 @@ describe('StoredCard', () => {
         expect(screen.queryByText('Expiry date', { exact: false })).toBeNull(); // non-presence
 
         expect(screen.getAllByText('Security code', { exact: true })).toBeTruthy();
-        expect(screen.getByRole('img', { name: 'Security code' })).toBeTruthy();
+        expect(screen.getByRole('img', { name: 'Security code 3 digits on back of card' })).toBeTruthy();
     });
 
     test('Renders a StoredCard field, without expiryDate (relevant data is empty string); and with cvc field', () => {
@@ -62,7 +62,7 @@ describe('StoredCard', () => {
         expect(screen.queryByText('Expiry date', { exact: false })).toBeNull(); // non-presence
 
         expect(screen.getAllByText('Security code', { exact: true })).toBeTruthy();
-        expect(screen.getByRole('img', { name: 'Security code' })).toBeTruthy();
+        expect(screen.getByRole('img', { name: 'Security code 3 digits on back of card' })).toBeTruthy();
     });
 
     test('Renders a StoredCard field, without expiryDate (relevant data is missing); and with cvc field', () => {
@@ -75,6 +75,6 @@ describe('StoredCard', () => {
         expect(screen.queryByText('Expiry date', { exact: false })).toBeNull(); // non-presence
 
         expect(screen.getAllByText('Security code', { exact: true })).toBeTruthy();
-        expect(screen.getByRole('img', { name: 'Security code' })).toBeTruthy();
+        expect(screen.getByRole('img', { name: 'Security code 3 digits on back of card' })).toBeTruthy();
     });
 });