+++ packages/lib/storybook/config/commonConfig.ts
@@ -1,4 +1,4 @@
-const protocol = window.location.protocol; // gives 'http:' or 'https:' i.e. adds the colon
+export const protocol = window.location.protocol; // gives 'http:' or 'https:' i.e. adds the colon
 export const DEFAULT_SHOPPER_LOCALE = 'en-US';
 export const DEFAULT_COUNTRY_CODE = 'US';
 export const DEFAULT_AMOUNT_VALUE = 25900;